everything working except color, demo in zGameTest

+++ math/internal/units/conversions.c
@@ -2,9 +2,9 @@
 #include <math.h>
 
 float OCT_deg2rad(float degrees) {
-	return degrees * (M_PI / 180.0f);
+	return degrees * ((float)M_PI / 180.0f);
 }
 
 float OCT_rad2deg(float radians) {
-	return radians * (180.0f / M_PI);
+	return radians * (180.0f / (float)M_PI);
 }
